requirements.txt
pillow==11.2.1
psutil==7.0.0
pystray==0.19.5
python-xlib==0.33
six==1.17.0
